Qt text search crash fix
About 5 weeks ago, commit a97ff6234ea222dba34cdd8df72038265cb9e871
changed Qt's searching in text windows to be able to find a match
on the very first line.  It assumed that the first line would be
the current line except when repeating the same search after a
match.  But after a failed search the current line index is -1 and
starting a new search would crash trying to look that line up.