Qt text search crash fix
About 5 weeks ago, commit c205081354f8fd3e88b9783eaedc47b7c0371f6e
changed Qt's searching in text windows to be able to find a match
on the very first line.  It assumed that the first line would be
the current line except when repeating the same search after a
match.  But after a failed search the current line index is -1 and
starting a new search would crash trying to look that line up.