^G prompting revisited
CP_TRYLIM-1 was the right value when the prompt augmentation
was at the top of the loop before the first prompt, but should
been changed to CP_TRYLIM when that got moved to the bottom of
the loop.

First prompt:
|Create what kind of monster?
Second and subsequent prompts if first attempt is unsuccessful:
|Create what kind of monster? [type name or symbol]

Prior to this fix, the shorter prompt was being used on the
first and second tries and not switching to the longer one until
the third.