wizard mode wishing for doors
It was possible to wish for a secret door, if done at a wall or
door location, but not for a regular door.  Add that.  (Dig
followed by locking magic possibly followed by open or by kick
could cover most of the details without wish support, but there
wasn't any way to force a closed or locked door to be trapped.)

The wish request can include "trapped" (or "untrapped", the
default) and/or one of the 5 door states: "locked", "closed",
"open", "broken", or "doorless" (default is "closed").  If more
than one state is specified, the last one in the wish text
overrides others.  If trapped is specified with open or broken or
doorless, it will be ignored.

Allow "looted" when wishing for a fountain, sink, throne, or tree.
For the ones with multiple loot tracking bits, it sets them all.

Explicitly reject a wish for a wall rather than claiming nothing
of that description exists.

Formatting:  wrap various wide lines.